requirements.txt
# Core dependencies
rich>=13.0.0
pydantic>=1.10.0,<2.0.0
PyYAML>=6.0
pandas>=1.5.0
numpy>=1.21.0
yfinance>=0.2.0

# Data handling
pyarrow>=10.0.0

# Testing
pytest>=7.0.0

# Development
black>=22.0.0
isort>=5.10.0
flake8>=5.0.0